Fix routing so that links actually work
