+++ src/snapshots-app/client/bundles/pages/Course.jsx
@@ -1,15 +1,16 @@
 import React from "react";
 
 import { useParams } from "react-router";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { TextField, Box } from "@mui/material";
 import { DataGrid } from "@mui/x-data-grid";
 import { styled } from "@mui/material/styles";
-import { useAtom, useAtomValue } from "jotai";
+import { useAtomValue } from "jotai";
 
-import { coursesAtom, assignmentsAtom } from "../state/atoms";
+import { assignmentsAtom } from "../state/atoms";
 import TableCellNavLink from "../components/common/TableCellNavLink";
 
+// TODO do below and rename submission layout to timeline and move files around
 // TODO: rename paths and components to be consistent
 function AssignmentsTable({ courseId, assignments }) {
   const [search, setSearch] = useState("");
@@ -95,34 +96,12 @@ function AssignmentsTable({ courseId, assignments }) {
 
 function Course() {
   const routeParams = useParams();
-  const courses = useAtomValue(coursesAtom);
-  const selectedCourse = courses.find(
-    (course) => course.id.toString() === routeParams.courseId,
-  );
-  const [assignments, setAssignments] = useAtom(assignmentsAtom);
-
-  useEffect(() => {
-    fetch(`/api/assignments/${selectedCourse.id}`, {
-      method: "GET",
-    })
-      .then((response) => {
-        if (!response.ok) {
-          throw new Error(`HTTP error! Status: ${response.status}`);
-        }
-        return response.json();
-      })
-      .then((responseData) => {
-        setAssignments(responseData["assignments"]);
-      })
-      .catch((error) => {
-        throw new Error(`HTTP error! Error: ${error}`);
-      });
-  }, [selectedCourse, setAssignments]);
+  const assignments = useAtomValue(assignmentsAtom);
 
   return (
     <div style={{ paddingLeft: "1rem", paddingRight: "1rem" }}>
       <AssignmentsTable
-        courseId={selectedCourse.id}
+        courseId={routeParams.courseId}
         assignments={assignments}
       />
     </div>
